+++ NEWS.md
@@ -6,6 +6,10 @@
 - Added linear momentum lost by gravitational waves along the x and y directions
   (starting from a contribution by @konrad-topolski)
 
+#### Bug fixes
+
+- Added error in `_plot_horizon_on_plane` when the horizon cannot be plotted
+
 ## Version 1.3.3 (19 November 2021)
 
 #### Bug fixes